refactor: 레이어드 아키텍처 적용
