+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -7,38 +7,32 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dao.ReservationTimeDao;
-import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
+import roomescape.service.ReservationTimeService;
 
 @RestController
 public class ReservationTimeController {
 
-    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationTimeService reservationTimeService;
 
-    public ReservationTimeController(ReservationTimeDao reservationTimeDao) {
-        this.reservationTimeDao = reservationTimeDao;
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
     }
 
     @GetMapping("/times")
     public List<ReservationTimeResponseDto> readReservationTimes() {
-        List<ReservationTime> reservationTimes = reservationTimeDao.findAllReservationTimes();
-        return reservationTimes.stream()
-            .map(ReservationTimeResponseDto::from)
-            .toList();
+        return reservationTimeService.getAllReservationTimes();
     }
 
     @PostMapping("/times")
     public ReservationTimeResponseDto saveReservationTime(
         @RequestBody ReservationTimeRequestDto reservationTimeRequestDto) {
-        ReservationTime reservationTime = reservationTimeRequestDto.toReservationTime();
-        reservationTimeDao.saveReservationTime(reservationTime);
-        return ReservationTimeResponseDto.from(reservationTime);
+        return reservationTimeService.saveReservationTime(reservationTimeRequestDto);
     }
 
     @DeleteMapping("/times/{id}")
     public void deleteReservationTime(@PathVariable(name = "id") Long id) {
-        reservationTimeDao.deleteReservationTime(id);
+        reservationTimeService.deleteReservationTime(id);
     }
 }
