+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,0 +1,36 @@
+package roomescape.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.ReservationTimeRequestDto;
+import roomescape.dto.ReservationTimeResponseDto;
+
+@Service
+public class ReservationTimeService {
+
+    private final ReservationTimeDao reservationTimeDao;
+
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
+    }
+
+    public List<ReservationTimeResponseDto> getAllReservationTimes() {
+        List<ReservationTime> reservationTimes = reservationTimeDao.findAllReservationTimes();
+        return reservationTimes.stream()
+            .map(ReservationTimeResponseDto::from)
+            .toList();
+    }
+
+    public ReservationTimeResponseDto saveReservationTime(
+        ReservationTimeRequestDto reservationTimeRequestDto) {
+        ReservationTime reservationTime = reservationTimeRequestDto.toReservationTime();
+        reservationTimeDao.saveReservationTime(reservationTime);
+        return ReservationTimeResponseDto.from(reservationTime);
+    }
+
+    public void deleteReservationTime(Long id) {
+        reservationTimeDao.deleteReservationTime(id);
+    }
+}
